pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
